+++ src/jigsaw/component/box/box.type.ts
@@ -1,0 +1,7 @@
+import {EventEmitter} from "@angular/core";
+import {JigsawEditableBox} from "./editable-box";
+
+export interface JigsawEditableBoxShell {
+    element: HTMLElement;
+    fill: EventEmitter<JigsawEditableBox>;
+}
