+++ src/jigsaw/component/box/editable-box.scss
