+++ src/jigsaw/component/view-editor/view-editor.ts
@@ -1,284 +1,10 @@
-import {
-    AfterViewInit, Component, ComponentFactoryResolver, ComponentRef,
-    ElementRef, EmbeddedViewRef, EventEmitter,
-    Input, NgModule, OnDestroy, OnInit, Output, QueryList,
-    Renderer2, TemplateRef, Type, ViewChild, ViewChildren
-} from "@angular/core";
+import {Component, ElementRef, EventEmitter, Input, NgModule, Output} from "@angular/core";
 import {LayoutData} from "../../core/data/layout-data";
 import {CommonModule} from "@angular/common";
-import {AbstractJigsawComponent, JigsawCommonModule, JigsawRendererHost} from "../common";
-import {JigsawBoxResizableBase} from "../box/box.common";
-import {CallbackRemoval} from "../../core/utils/common-utils";
-import {ComponentInput, ComponentMetaData} from "./view-editor.type";
+import {AbstractJigsawComponent, JigsawCommonModule} from "../common";
 import {JigsawResizableModule} from "../../directive/resizable/resizable";
-
-@Component({
-    selector: 'jigsaw-editable-box, j-editable-box',
-    templateUrl: './editable-box.html',
-    host: {
-        '[class.jigsaw-editable-box]': 'true',
-        '[class.jigsaw-box]': 'true',
-        '[class.jigsaw-flex]': 'type == "flex"',
-        '[style.width]': 'width',
-        '[style.height]': 'height',
-    }
-})
-export class JigsawEditableBox extends JigsawBoxResizableBase implements AfterViewInit, OnDestroy, OnInit {
-    private _removeElementScrollEvent: CallbackRemoval;
-    private _removeDataRefreshListener: CallbackRemoval;
-
-    constructor(elementRef: ElementRef,
-                renderer: Renderer2,
-                private _componentFactoryResolver: ComponentFactoryResolver) {
-        super(elementRef, renderer);
-    }
-
-    private _data: LayoutData;
-
-    @Input()
-    public get data() {
-        return this._data;
-    }
-
-    public set data(value: LayoutData) {
-        if (!(value instanceof LayoutData)) return;
-        this._rendererHost.viewContainerRef.clear();
-        this._data = value;
-        if (this._removeDataRefreshListener) {
-            this._removeDataRefreshListener();
-        }
-        this._removeDataRefreshListener = this._data.onRefresh(() => {
-        })
-    }
-
-    @Output()
-    public dataChange = new EventEmitter<LayoutData>();
-
-    @ViewChildren(JigsawEditableBox)
-    protected childrenBox: QueryList<JigsawEditableBox>;
-
-    @Output()
-    public directionChange = new EventEmitter<string>();
-
-    @Output()
-    public growChange = new EventEmitter<number>();
-
-    @Output()
-    public remove = new EventEmitter<LayoutData>();
-
-    @ViewChild(JigsawRendererHost)
-    private _rendererHost: JigsawRendererHost;
-
-    @Input()
-    public editable: boolean;
-
-    @Input()
-    public blocked: boolean;
-
-    @Input()
-    public isFirst: boolean;
-
-    @Input()
-    public parent: JigsawEditableBox;
-
-    @Input()
-    public resizeLineWidth: string;
-
-    @Input()
-    public parentViewEditor: JigsawViewEditor;
-
-    /**
-     * @internal
-     */
-    public _$addItems(direction: string) {
-        this._rendererHost.viewContainerRef.clear();
-        if (this._removeElementScrollEvent) {
-            this._removeElementScrollEvent();
-        }
-        if (!this.data) {
-            this.data = new LayoutData;
-            this.dataChange.emit(this.data);
-        }
-        this.direction = direction;
-        this.directionChange.emit(this.direction);
-        // 目前先平分创建两个node
-        this.data.nodes = [new LayoutData, new LayoutData];
-        // 内容信息搬到第一个子节点
-        this.data.nodes[0].componentMetaDataList = this.data.componentMetaDataList;
-        this.data.nodes[0].innerHtml = this.data.innerHtml;
-        // 重置当前内容信息
-        this.data.componentMetaDataList = [];
-        this.data.innerHtml = ''
-    }
-
-    /**
-     * @internal
-     */
-    public _$remove() {
-        this.remove.emit(this.data);
-    }
-
-    /**
-     * @internal
-     */
-    public _$removeItem(item: LayoutData) {
-        if (!this.data || !(this.data.nodes instanceof Array)) return;
-
-        const index = this.data.nodes.findIndex(node => node === item);
-        this.data.nodes.splice(index, 1);
-        if (this.data.nodes.length == 1) {
-            // 处理多余的box
-            const node = this.data.nodes[0];
-            if (node.nodes instanceof Array && node.nodes.length > 0) {
-                // 拿取nodes
-                this.data.nodes = node.nodes;
-                this.direction = node.direction ? node.direction : 'h'; // 默认是'h'
-            } else {
-                // 拿取内容
-                this.data.componentMetaDataList = node.componentMetaDataList;
-                this.data.innerHtml = node.innerHtml;
-                this.data.nodes = [];
-                this.direction = null;
-                this._renderComponents(this.data.componentMetaDataList);
-                setTimeout(() => {
-                    // 等待 option bar & block 渲染
-                    this._bindScrollEvent();
-                })
-            }
-            this.directionChange.emit(this.direction);
-        }
-    }
-
-    /**
-     * @internal
-     */
-    public _$addContent() {
-        this.parentViewEditor.fill.emit(this);
-    }
-
-    public addContent(componentMetaDataList: ComponentMetaData[]) {
-        if (!this.data) {
-            return;
-        }
-        this._renderComponents(componentMetaDataList);
-        this.data.setComponentMetaData(componentMetaDataList);
-        this._bindScrollEvent();
-    }
-
-    /**
-     * 根据componentMetaDataList信息在box里面渲染需要的组件
-     * @param {ComponentMetaData[]} componentMetaDataList
-     * @private
-     */
-    private _renderComponents(componentMetaDataList: ComponentMetaData[]) {
-        if (!(componentMetaDataList instanceof Array) || componentMetaDataList.length == 0) return;
-        this._rendererHost.viewContainerRef.clear();
-        this.data.components = []; // 初始化
-        componentMetaDataList.forEach(componentMetaData => {
-            this.data.components.push(this._rendererFactory(componentMetaData.component, componentMetaData.inputs));
-        });
-    }
-
-    private _rendererFactory(renderer: Type<any> | TemplateRef<any>, inputs: ComponentInput[]): ComponentRef<any> | EmbeddedViewRef<any> {
-        if (renderer instanceof TemplateRef) {
-            const context = {};
-            if (inputs) {
-                inputs.forEach(input => context[input.property] = input.binding);
-            }
-            return this._rendererHost.viewContainerRef.createEmbeddedView(renderer, {
-                context: context
-            });
-        } else if (renderer) {
-            let componentFactory = this._componentFactoryResolver.resolveComponentFactory(renderer);
-            let componentRef = this._rendererHost.viewContainerRef.createComponent(componentFactory);
-            if (inputs) {
-                inputs.forEach(input => componentRef.instance[input.property] = input.default);
-            }
-            return componentRef;
-        }
-        return null;
-    }
-
-    /**
-     * 绑定box的滚动监听，处理option bar & block的位置问题
-     * @private
-     */
-    private _bindScrollEvent() {
-        // 有node不绑定scroll
-        if (this.data && this.data.nodes instanceof Array && this.data.nodes.length > 0) return;
-
-        const block = this.element.querySelector('.jigsaw-editable-box-block');
-        const optionBox = this.element.querySelector('.jigsaw-editable-box-option-box');
-        const optionBar = this.element.querySelector('.jigsaw-editable-box-option-bar');
-        const resizeBar = this.element.querySelector('.jigsaw-editable-box-resize');
-        if (this._removeElementScrollEvent) {
-            this._removeElementScrollEvent();
-        }
-        this._removeElementScrollEvent = this._renderer.listen(this.element, 'scroll', () => {
-            if (block) {
-                this._renderer.setStyle(block, 'top', this.element.scrollTop + 'px');
-                this._renderer.setStyle(block, 'left', this.element.scrollLeft + 'px');
-            }
-            if (optionBox) {
-                this._renderer.setStyle(optionBox, 'top', this.element.scrollTop ? (this.element.offsetHeight / 2 + this.element.scrollTop + 'px') : '50%');
-                this._renderer.setStyle(optionBox, 'left', this.element.scrollLeft ? (this.element.offsetWidth / 2 + this.element.scrollLeft + 'px') : '50%');
-            }
-            if (optionBar) {
-                this._renderer.setStyle(optionBar, 'top', this.element.scrollTop ? (this.element.offsetHeight / 2 + this.element.scrollTop + 'px') : '50%');
-                this._renderer.setStyle(optionBar, 'left', this.element.scrollLeft ? (this.element.offsetWidth / 2 + this.element.scrollLeft + 'px') : '50%');
-            }
-            if (resizeBar) {
-                this._renderer.setStyle(resizeBar, 'top', this.element.scrollTop + 'px');
-                this._renderer.setStyle(resizeBar, 'left', this.element.scrollLeft + 'px');
-            }
-        })
-    }
-
-    @ViewChild('resizeLine')
-    public resizeLine: ElementRef;
-
-    @ViewChild('resizingLine')
-    public resizingLine: ElementRef;
-
-    /**
-     * @internal
-     */
-    public _$handleResizeStart(event) {
-        super._$handleResizeStart(event);
-        this._renderer.addClass(this.parentViewEditor.element, 'jigsaw-view-editor-resizing');
-    }
-
-    /**
-     * @internal
-     */
-    public _$handleResizeEnd() {
-        this._renderer.removeClass(this.parentViewEditor.element, 'jigsaw-view-editor-resizing');
-    }
-
-    ngOnInit() {
-        super.ngOnInit();
-        if (this.data) {
-            this._renderComponents(this.data.componentMetaDataList);
-            this.data.box = this;
-        }
-    }
-
-    ngAfterViewInit() {
-        this.checkFlex();
-        // 等待 option bar & block 渲染
-        this._bindScrollEvent();
-    }
-
-    ngOnDestroy() {
-        this._rendererHost.viewContainerRef.clear();
-        if (this._removeElementScrollEvent) {
-            this._removeElementScrollEvent();
-        }
-        if (this._removeDataRefreshListener) {
-            this._removeDataRefreshListener();
-        }
-    }
-}
+import {JigsawEditableBoxShell} from "../box/box.type";
+import {JigsawEditableBox} from "../box/editable-box";
 
 @Component({
     selector: 'jigsaw-view-editor, j-view-editor',
@@ -294,7 +20,7 @@ export class JigsawEditableBox extends JigsawBoxResizableBase implements AfterVi
         '[style.height]': 'height',
     }
 })
-export class JigsawViewEditor extends AbstractJigsawComponent {
+export class JigsawViewEditor extends AbstractJigsawComponent implements JigsawEditableBoxShell {
     public element: HTMLElement;
 
     constructor(elementRef: ElementRef) {
