+++ src/jigsaw/core/theming/all-theme.scss
@@ -39,7 +39,7 @@
 @import "../../component/viewport/viewport";
 @import "../../component/scrollbar/scrollbar";
 @import "../../component/box/box";
-@import "../../component/view-editor/editable-box";
+@import "../../component/box/editable-box";
